Cart Changes
Cart Controller change, CartRepo change,  ShoppingCart Index change

+++ BookCave/Repositories/CartRepo.cs
@@ -25,16 +25,42 @@ namespace BookCave.Repositories
 
         public void AddBookToCart(int bookId, string userId)
         {
-            var BookToCart = new Cart{ UserId = userId, BookId = bookId };
-            _db.Carts.Add(BookToCart);
-            _db.SaveChanges();
+            bool isInCart = _db.Carts.Any(u => u.UserId == userId && u.BookId == bookId);
+            if(isInCart)
+            {
+                var Cart = _db.Carts.Where(u => u.UserId == userId && u.BookId == bookId).SingleOrDefault();
+                Cart.Quantity += 1;
+                _db.SaveChanges();
+            }
+            else 
+            {
+                var BookToCart = new Cart{ UserId = userId, BookId = bookId, Quantity = 1};
+                _db.Carts.Add(BookToCart);
+                _db.SaveChanges();                
+            } 
+
         }
 
         public void RemoveFromCart(int bookId)
         {
             var BookToRemove = _db.Carts.Where(c => c.BookId == bookId).FirstOrDefault();
-            _db.Remove(BookToRemove);
-            _db.SaveChanges();
+            if(BookToRemove.Quantity > 1)
+            {
+                BookToRemove.Quantity -= 1;
+                _db.SaveChanges();
+            }
+            else if(BookToRemove.Quantity == 1)
+            {
+                _db.Remove(BookToRemove);
+                _db.SaveChanges();
+            }
+            
+        }
+        public void IncreaseInCart(int bookId)
+        {
+            var BookToIncrease = _db.Carts.Where(c => c.BookId == bookId).FirstOrDefault();
+            BookToIncrease.Quantity += 1;
+             _db.SaveChanges();
         }
 
         public List<CartViewModel> GetCart(string userId)
@@ -45,8 +71,10 @@ namespace BookCave.Repositories
                         select new CartViewModel
                         {
                             Id = b.Id,
+                            UserId = c.UserId,
                             Title = b.Title,
-                            Price = b.Price 
+                            Price = b.Price,
+                            Quantity = c.Quantity,
                         }).ToList();
             return Cart;
         }
